Separate refresh endpoint configuration into its own auto config.  Fixes #151
