+++ spring-cloud-bus/src/main/java/org/springframework/cloud/bus/BusAutoConfiguration.java
@@ -29,7 +29,6 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.cloud.bus.endpoint.EnvironmentBusEndpoint;
-import org.springframework.cloud.bus.endpoint.RefreshBusEndpoint;
 import org.springframework.cloud.bus.event.AckRemoteApplicationEvent;
 import org.springframework.cloud.bus.event.EnvironmentChangeListener;
 import org.springframework.cloud.bus.event.RefreshListener;
@@ -38,7 +37,6 @@ import org.springframework.cloud.bus.event.SentApplicationEvent;
 import org.springframework.cloud.bus.event.TraceListener;
 import org.springframework.cloud.context.environment.EnvironmentManager;
 import org.springframework.cloud.context.refresh.ContextRefresher;
-import org.springframework.cloud.context.scope.refresh.RefreshScope;
 import org.springframework.cloud.stream.annotation.EnableBinding;
 import org.springframework.cloud.stream.annotation.Output;
 import org.springframework.cloud.stream.annotation.StreamListener;
@@ -172,30 +170,6 @@ public class BusAutoConfiguration implements ApplicationEventPublisherAware {
 				BusProperties properties) {
 			ServiceMatcher serviceMatcher = new ServiceMatcher(pathMatcher, properties.getId());
 			return serviceMatcher;
-		}
-
-	}
-
-	@Bean
-	@ConditionalOnProperty(value = "spring.cloud.bus.refresh.enabled", matchIfMissing = true)
-	@ConditionalOnBean(ContextRefresher.class)
-	public RefreshListener refreshListener(ContextRefresher contextRefresher) {
-		return new RefreshListener(contextRefresher);
-	}
-
-	@Configuration
-	@ConditionalOnClass({ Endpoint.class, RefreshScope.class })
-	protected static class BusRefreshConfiguration {
-
-		@Configuration
-		@ConditionalOnBean(ContextRefresher.class)
-		protected static class BusRefreshEndpointConfiguration {
-			@Bean
-			@ConditionalOnEnabledEndpoint
-			public RefreshBusEndpoint refreshBusEndpoint(ApplicationContext context,
-					BusProperties bus) {
-				return new RefreshBusEndpoint(context, bus.getId());
-			}
 		}
 
 	}
