+++ spring-cloud-bus/src/test/java/org/springframework/cloud/bus/BusAutoConfigurationClassPathTests.java
@@ -22,7 +22,7 @@ public class BusAutoConfigurationClassPathTests {
     public void refreshListenerCreatedWithoutActuator() {
         new ApplicationContextRunner()
                 .withConfiguration(AutoConfigurations.of(RefreshAutoConfiguration.class,
-                        BusAutoConfiguration.class))
+                        BusRefreshAutoConfiguration.class))
                 .run(context -> assertThat(context)
                         .hasSingleBean(RefreshListener.class)
                         .doesNotHaveBean(RefreshBusEndpoint.class));
